bondy_broker_bridge behaviour (bondy_broker_bridge v0.1.0)

This module defines the behaviour for providing event bridging functionality, allowing a supervised process (implemented via bondy_subscriber) to consume WAMP events based on a normal subscription to publish (or produce) those events to an external system, e.g. another message broker, by previously applying a transformation specification based on a templating language..

Each broker bridge is implemented as a callback module exporting a predefined set of callback functions.

Callbacks

apply_action/1

Specs

apply_action(Action :: map()) -> ok | {retry, Reason :: any()} | {error, Reason :: any()}.

init/1

Specs

init(Config :: any()) -> {ok, Ctxt :: #{binary() => any()}} | {error, Reason :: any()}.

terminate/2

Specs

terminate(Reason :: any(), State :: any()) -> ok.

validate_action/1

Specs

validate_action(Action :: map()) -> {ok, ValidAction :: map()} | {error, Reason :: any()}.