bondy_broker_bridge_manager (bondy_broker_bridge v0.1.0)

This module provides event bridging functionality, allowing a supervised process (implemented via bondy_subscriber) to subscribe to WAMP events and process and/or forward those events to an external system, e.g. publish to another message broker.

A subscription can be created at runtime using the subscribe/5, or at system boot time by setting the application's config_file environment variable which should have the filename of a valid Broker Bridge Specification File.

Each broker bridge is implemented as a module implementing the bondy_broker_bridge behaviour.

## Action Specification Map.

## mops Evaluation Context

The mops context is map containing the following:

     erlang
  #{
      <<"broker">> => #{
          <<"node">> => binary()
          <<"agent">> => binary()
      },
      <<"event">> => #{
          <<"realm">> => uri(),
          <<"topic">> => uri(),
          <<"subscription_id">> => integer(),
          <<"publication_id">> => integer(),
          <<"details">> => map(), % WAMP EVENT.details
          <<"args">> => list(),
          <<"kwargs">> => map(),
          <<"ingestion_timestamp">> => integer()
      }
  }.

## Broker Bridge Specification File.

Example:

     json
  {
      "id":"com.leapsight.test",
      "meta":{},
      "subscriptions" : [
          {
              "bridge": "bondy_kafka_bridge",
              "match": {
                  "realm": "com.leapsight.test",
                  "topic" : "com.leapsight.example_event",
                  "options": {"match": "exact"}
              },
              "action": {
                  "type": "produce_sync",
                  "topic": "{{kafka.topics.wamp_events}}",
                  "key": "\"{{event.topic}}/{{event.publication_id}}\"",
                  "value": "{{event}}",
                  "options" : {
                      "client_id": "default",
                      "acknowledge": true,
                      "required_acks": "all",
                      "partition": null,
                      "partitioner": {
                          "algorithm": "fnv32a",
                          "value": "\"{{event.topic}}/{{event.publication_id}}\""
                      },
                      "encoding": "json"
                  }
              }
          }
      ]
  }

Types

bridge/0

Specs

bridge() :: map().

id/0

Specs

id() :: 0..9007199254740992.

subscription_detail/0

Specs

subscription_detail() :: map().

uri/0

Specs

uri() :: binary().

Functions

bridge(Mod)

Specs

bridge(module()) -> [bridge()].

Returns the bridge configuration identified by its module name.

bridges()

Specs

bridges() -> [bridge()].

Lists all the configured bridge configurations.

code_change(OldVsn, State, Extra)

handle_call(Event, From, State)

handle_cast(Event, State)

handle_continue(_, State0)

handle_info(Info, State)

init(_)

load(Term)

Specs

load(file:filename() | map()) -> ok | {error, invalid_specification_format | any()}.

Parses the provided Broker Bridge Specification and creates all the provided subscriptions.

start_link()

Internal function called by bondy_broker_bridge_sup.

subscribe(RealmUri, Opts, Topic, Bridge, Spec)

Specs

subscribe(uri(), map(), uri(), Bridge :: module(), Spec :: map()) ->
             {ok, id()} | {error, already_exists}.

Creates a subscription using bondy_broker. This results in a new supervised bondy_subscriber processed that subcribes to {Realm, Topic} and forwards any received publication (event) to the bridge identified by Bridge.
Returns the tuple {ok, Pid} where Pid is the pid() of the supervised process or the tuple {error, Reason}.

subscriptions(BridgeId)

Specs

subscriptions(BridgeId :: bridge()) -> [subscription_detail()].

terminate(Reason, State)

unsubscribe(Id)

Specs

unsubscribe(id()) -> ok | {error, not_found}.

validate_spec(Map)

Specs

validate_spec(map()) -> {ok, map()} | {error, any()}.