bondy_kafka_bridge (bondy_broker_bridge v0.1.0)

Functions

apply_action(Action)

Evaluates the action specification Action against the context Ctxt using mops and produces to Kafka.

init(Config)

Initialises the Kafka clients provided by the configuration.

terminate(Reason, State)

validate_action(Action0)

Validates the action specification. An action spec is a map containing the following keys:

* type :: binary() - <<"produce_sync">>. Optional, the default value is <<"produce_sync">>. * topic :: binary() - the Kafka topic we should produce to. * key :: binary() - the kafka message's key * value :: any() - the kafka message's value * options :: map() - a map containing the following keys

     erlang
  #{
      <<"type">> <<"produce">>,
      <<"topic": <<"com.magenta.wamp_events",
      <<"key": "\"{{event.topic}}/{{event.publication_id}}\"",
      <<"value": "{{event}}",
      <<"options" : {
          <<"client_id": "default",
          <<"acknowledge": true,
          <<"required_acks": "all",
          <<"partition": null,
          <<"partitioner": {
              "algorithm": "fnv32a",
              "value": "\"{{event.topic}}/{{event.publication_id}}\""
          },
          <<"encoding">>: <<"json">>
      }
  }